Refactor analytics endpoints to use course public_id instead of internal id, make BaseAnalyticsView abstract, and remove nullable user field from QuestionAttempt
