+++ analytics/urls.py
@@ -1,8 +1,8 @@
 from django.urls import path
 
-from .views import ClassAverageView
-from .views import TimePerQuestionView
-from .views import UnitDistributionView
+from .views.class_average import ClassAverageView
+from .views.time_per_question import TimePerQuestionView
+from .views.unit_distribution import UnitDistributionView
 
 urlpatterns = [
     path("class-averages/", ClassAverageView.as_view(), name="class_averages"),
